Add completion event to pcapper SNAPSHOT so that the caller (say, loghole) knows when the logfiles are actually rolled.
Event groups complicated things a bit. To make this work properly, we
no longer subscribe to the link-tracemon event, but instead use a
real event group, created by assign wrapper for all of the linktrace
agents. So, you can know do things like this:

	tevc -w -e testbed/TT now link0_tracemon snapshot
or
	tevc -w -e testbed/TT now __all_tracemon snapshot

where __all_tracemon is a group of all tracemon agents for all links and
lans. I plan to change loghole to use this.