Second attempt to fix the problem of duplicate log entries.  I am 99.99% sure this will get 100% of the cases, and 99.999% sure it won't break anything.
It basically detects when the DB handle is a child and if so set
"InaciveDestroy" before the database handle DESTROY method is called.
Since the DB handle can be closed in several different places I created a
new class to override the Db Handle (the Mysql class) DESTROY method. The
other alternative is to add special code anywhere where the database handle
could be destroyed which is when every a reconnect is done and when the
module exists.  The later would have involved putting code in the END block.
I think the new class method is simpler for that reason.


Also, add a note about patching Mysql.pm in doc/UPDATING.