Change the way vnode_setup handles plab nodes a bit to avoid a couple of buggy situations.
  * Don't try vnodesetup -h on plab nodes

This can hang, or even fail.  Since nothing useful is conveyed by this
step, just skip it, set the node's state to SHUTDOWN, and ask pl_conf
on the node to remove the vserver.

  * Set plab node's alloc state to TBDB_ALLOCSTATE_RES_INIT_DIRTY after
    instantiation.

This avoids a bug where Emulab cluster nodes fail to come up, and so
os_setup never waits on the plab vnodes (now that they are started in
parallel with physical node setup).  Previously their alloc state made
them look clean, and so the vservers would not be reaped during
teardown.