Add compression option to sync option of loghole. When turned on, any file greater the 512K is automatically compressed with gzip. Might need to make this number bigger; we shall see.
If you run emacs, put this in your .emacs file.

	(load "jka-compr")
	(jka-compr-install)

and any time you visit a file that ends in one of the standard compression
extensions, emacs will automatically do the uncompress for you on the data
in the buffer (not the actual disk file of course). Very convenient.

You can also get your browser to do the same, but I leave that as an
exercise for the reader.