Ick.  Add a retry loop for getting ifconfig info from TMCD.
This is needed for multiplexed links which use tagged VLANs as their
implementation.  For these links, it is possible at initial swapin time
that the VLAN tag has not been setup in the DB by the time we query it.
So if we get back a 0 tag, we sleep awhile and try again.