Partial fix to FS#137 concerning plab nodes with broken nameserver setups that cannot resolve even their own name.
Put a check in the generated rc.plab to see if the node can ping itself
(i.e., that its hostname resolves).  If not, we put an entry for ourselves
in /etc/hosts.

This does not solve the general problem of a bad nameserver, but allows
our infrastructure to completely setup.  Whether this is a step forward or
not is debatable, but it makes me feel all warm inside.