Update setup{,-ops}.txt for FreeBSD-6.1 .
The packagedir is now passed to {ops,boss}-install with a -p argument instead of
as a PKG_PATH envar.  If the old way is done, PKG_PATH is ignored and packages are
all built from sources.  This failed in many ways (due to sync problems between the
6.1 packages and ports tarballs?)  In any case, to prevent others falling down the
same rabbit-hole, building from sources is now off by default, requiring a -s
argument to be given to do it.  Libinstall now shows PhaseSkip reasons, to make
it more clear what's going on.

Also tweak dhcpd_makeconf to handle a null DHCPD_DYNRANGE in the defs file,
to prevent a bogus dhcpd response from an inner emulab, half set-up for testing.