Attempt to reconcile the definition of some program-agent environment variables on cluster and plab nodes.
NODECNET (new)
The fully-qualified (Emulab) name of the node this program agent is running on.
This name resolves to the IP address of the control network interface
of the node.

NODECNETIP (new)
The IP address of the control network interface.
This address should not be advertised to, or used by, applications
within an experiment as it will cause all traffic to flow over the control
network rather than the experimental network.

NODE
The unqualified name of the node this program agent is running on.
For nodes with experimental interfaces,
this name resolves to the IP address of an experimental interface on the node.
For nodes with more than one experimental interface,
there is no guarantee which one it will resolve to.
For nodes with no experimental interfaces,
the name will not resolve.

NODEIP
The IP address of the experiment network interface that NODE resolves to.
For nodes with no experimental interfaces, this variable will not be set.