Modify LOADINFO command to include ACPI info for FreeBSD images.
node_type_attributes can now include a row with attrkey=use_acpi which can
either be set to:

  '1', if the FreeBSD image absolutely must use ACPI on this node_type
       in order to boot (e.g., our Dell SC440 machine)

  '0', if the FreeBSD image absolutely must NOT use ACPI on the node_type
       in order to boot (e.g., our pc2000s, possibly due to the IXP cards).

Otherwise, just don't set this attribute.

This value is queried by rc.frisbee and passed to slicefix after the disk
is loaded so that it can change /boot/loader.conf to enable/disable ACPI.

Note that ACPI (or lack thereof) has never been an issue with Linux or
Windows, so this is only meaningful to FreeBSD >= 5.0.

Note also that this does nothing for MFSes.  On machines that must use ACPI,
you need to use a special MFS for that node type (set via 'adminmfs_osid'
and 'diskloadmfs_osid' node_type_attributes).