In remote mode, don't limit the number of attempts to connect to the server.
This causes program agents to become isolated and marginalized, in short,
unresponsive to the needs of Emulab.  It is now a Loyal Emulab Soldier,
attempting reconnects til the end of time if necessary.

I left in the "keep alive" (idle timeout) since it doesn't seem to be causing
any problems.