New front page text, for the first time in a few years. More clearly enumerates the different experimental environments we support. Most of the new text only gets displayed at Utah, since no one else has wireless nodes, etc. Other sites get some short generic text.
New banner, plus other visual tweaks. The new banner is turned on via
the new @FANCYBANNER@ autoconf variable. This is turned on for
TBMAINSITE, but defaults to off for other sites. This is so that
existing sites which already have their own versions of the old banner
don't have to update them right away.

Made the usage iframe a little less prominent, by darkening it, and
making it ever so slightly transparent on browsers that support it.

Some minor visual tweaks to the background and content area.

Added specific IDs for the main menu subgroups so that if we want, we
can style them differently.

Man, IE is a pain in the ass.