sshhttp.py provides a means of speaking XMLRPC over HTTP using python file objects, not sockets.  The reason for doing this is because I expected to be talking to the new PLC4-era NM over an ssh tunnel, which would only allow me to talk to the NM over stdin/stdout (i.e., no ssh socket tunnels because there's only a pseudo shell on the NM side that proxies stdin/stdout to the NM).
I was trying not to have to reimplement the whole httplib, so I overrode
the bits that can't stand to use anything but a socket.  Lots of those.
Anyway, seems to work... so anytime you need to talk XMLRPC over HTTP but
not over a socket, this will work for you.  You'll just want to write your
own "real transport" method ... I only wrote one that popens a command.