Hackery to make sure that Plab slices always have boss/ops in their hosts file: we return info for boss/ops in the "hostnames" command, but only if the command would have otherwise returned something.
This is because the default hosts file we distribute in the rootball already
has boss/ops in it.  But if, during bootup, tmcd returns hostname info, that
hosts file gets overwritten and boss/ops info gets lost.

I could just grep out the info from the original hosts file and transfer it
to the one we are building, but what do I grep for: "boss", "ops",
"emulab.net"?  Don't really want to hardwire any of those in the client-side
script.

By returning this info via tmcd, I also don't have to modify the client-side
and thus don't need to build a new rootball!