A change for Dave that we discussed a couple of months back and yesterday Dave said he needs it now.
Currently the accounts that are returned to ron nodes are determined
by looking for the node type in the pcremote_ok slot of the projects
table. There is no per-user subgroup setting; everyone gets the group
of their project.

Dave needs to be able to assign remote users to specific subgroups
within a variety of projects, and I did not want to hack up the
pcremote_ok mechanism any further. So, new way to do this, that might
replace pcremote_ok at some point, but for now will just override it.

I am using the node_attributes table to drive what groups get
installed, and thus what users get accounts and a grouplist.

	inset into node_attributes values
	           ('ronXXX', "dp_projects", "10000,10001,...");

where the group list is specified as a list of gid_idx's. This works
out very nicely cause I can use a subquery and FIND_IN_SET clause, and
so the changes to tmcd where actually pretty easy.