Fix FS#151  "$ns report" get's called in seemingly random places"
It turns out not to be random. Instead send_report is being called
when something fails in a timeline. The intent here (I assume) is to
inform the user of the error. Unfortunately this has several
undesirable side-effects such as creating an unnecessary zip file and
overwritten the existing contents of report.mail. I fixed this by
creating a new function send_error_report which is like send report
but doesn't have the undesirable side-effects.