First step is allowing a different "standard" disk layout:
 * Somewhere in the DB, probably images, we will store the MBR "version"
   required by a partition (non-whole-disk) image.  This info will be
   passed to the client (us).

 * We use this info to install the correct MBR on the disk before starting
   frisbee.  The various MBRs (all two of them) will be stored in the
   frisbee MFS since they are only 512 bytes each.

This checkin handles the client-side of things and is backward compatible
(i.e., assumes MBR version 1 if nothing is passed).

One immediate benefit is that this will allow us to install a partition
image on a new (or otherwise MBR-free) disk, as it detects the case of no
MBR and installs one.