SanityCheck.py is the script for checking cross-correlation of Flexlab runs. It works for both TCP and UDP transfers. I also changed auto-pelab.ns and added a "-t" option to start-experiment and stop-experiment to enable these sanity checks.
Note: I haven't figured out how to kill sudo tcpdump sessions
on PlanetLab with program agents - so, for now they have to be
killed manually after a stop-experiment. Will change this after
talking to Mike.

The cross-correlation code works correctly - I may need to tweak it
a bit in the near future for averaging times.
Also, it puts too much load on Ops and (i think) takes too much time.
I will change it so that all the correlation calculations are done on
the experimental nodes.