Changes to switch from using tag to rtag. Note that this commit requires a locally modifed version of CVS that does a couple of extra things. The hacks are quite simple and currently on users. Briefly, there is a new config file option to specify a program to run when the tag operation is complete (similar to a module hook for rtag), and in tag.c I run that hook.
Check out a sandbox as before:

  cvs -d ops.emulab.net:/proj/$pid/templates/XXXXX/cvsrepo rtag mytag XXXXX

To create a new template (along the trunk), use the tag operation instead:

	cd sandbox
	cvs tag mtag

and what for the email that says the template is created. You can
continue to work along the trunk in this manner, although note that
the .template file is changing (by the backend commit), and you might
find it less distriacting to do a cvs update each time.

To work along a branch:

	cd sandbox
	cvs tag -r existing_tag -b newtag1
        cvs update -r newtag
        make changes and commits
	cvs tag newtag2

When working along branched you will want to do a cvs update to get
the .template file in sync.