The CVS repo will now have two subdirs, one called "setup" which is where the current stuff goes (datastore, nsfile, etc.). The other subdir is called "records" and will be populated with stuff that we currently put into the archive (the contents of the experiment directory).
There is a record for each run, currently named XXX,YY where XXX is
the instance ID and YY is the run number. I plan to add an XML file to
map these to the symbolic names, but these are not required to be
unique, which is why the directories are not named with them.