Add IP and privkey slots (key INDEX slots) to widearea_nodeinfo table. This will superceed the widearea_privkeys table that was used for RON/Netbed but was never done well.
The privkey is just something we generate and store on the dongle (as
a simplification, we just create a dongle per node instead of
generating them on the fly) since there are not very many of these
widearea (CMU) nodes. We store the IP here for convenience (there is
an interfaces table entry of course).

The IP is updated when the node tells us it has changed, using the
privkey to uniquely identify it (for checkin).