New widearea node checkin stuff for CMU. This stuff is quite a bit different then the original widearea code. Simpler, less dynamic.
First off, the wanodecreate script creates a new widearea_nodeinfo
entry.  These are nodes that will later checkin and be created as a
real node.  The input is a little xml file that you can use to specify
the stuff in the table entry (city, state, zip, etc). You can also
provide a privkey (no more then 64 chars), or one will be generated
for you.  For each one of these, create a Dongle Boot and stash the
privkey as /etc/emulab/emulab-privkey on the dongle. You do not assign
the IP address; the node will tell us that when it checks in.

A node checks in like this:

	https://$bossname/wanodecheckin.php?IP=$IP&privkey=$privkey
               &hostname=$hostname

The web page is simply a stub that makes sure the arguments don't have
any illegal characters, and then passes off to the backend.

The backend script checks the privkey and finds the widearea_nodeinfo.
The first time the node checks in, the node is created (db/Node.pm)
(nodes table, interfaces table, etc), and the node is moved to hwdown.
Subsequent checkins watch for changes to the IP or hostname, and issue
named_setup calls as needed.