First crack at limiting the node types assigned to new images.
Currently we declare the valid types for an image to be those that are of
class "pc" and for which we have at least one node in the testbed.

With this change, when an image is created as a snapshot from an existing
node, we look at the partition info for the existing image to further
restrict the types that the derived image will work on.

This will prevent images derived from old images like RHL90 or FBSD410
from getting newer node types such as pc745d as a valid type.

One problem with the current solution is that, we don't know til the form
is submitted once, what node you are snapshot'ing, so the first time a
user submits the form, it is almost certain to come back with an error.

Further, they have no choice but to uncheck the offending types.  If part
of their customization was to add support for a new node type...too bad,
ya' can't add it!  There are numerous ways to fix this, all of which exceed
my Friday afternoon mad hacking skilz.