Moderate rework of the Emulab UI. When logged in, all of the main menus that used to be on the left side, are now drop down menus along the top. The banner across the top will also be smaller and more space efficient.
The primary motivation for these changes is to leave more room for
actual content in the main window.

Tested with Firefox (1.5 and 2.0), Safari (2.0 and 3.0) and on
Internet Explorer Version 7. IE6 has some visual issues that might
make it a little annoying, but should be viewed as incentive to
upgrade to IE7 since it has been out a long time.