Convert the usage box at the top from an iframe to an inline table so that it can be actively clicked and toggled between the short status view and the long freenode view. There was no way to do this when it was an iframe, and the upside of not using an iframe is that iframes often mess up rendering and cause many cross-browser problems.
Since its not an iframe, I have to hook into the Sajax stuff even on
pages that already use Sajax, and I want to go to a different URL then
the current page (want to use currentusage.php3). So I tweaked the Sajax
stuff a bit, and roll out some inline sajax code in PAGEFOOTER().

Also, at Rob's request I added a cookie to remember the state of the
user's choice so that subsequent page loads get the same view. This
is done with an inline cookie in the javascript (document.cookies).