Fix for fixed interfaces: move the code that filters interfaces whose name's don't match from resolve_link() to find_best_link() - the latter returns only one link of each type, so it's too late to find the right interface by the time we get back out to resolve_link(). Also fixed a built-in assmption about the 'direction' that physical links go in. The old code (explicity) assumed that links to from node to switch - now we flip around our comparison if the directions of the links do not match.
This was a bugger, even though it didn't involve much code, since the
'flipping' involves a 'parity' issue - it's really easy to mess up the
parity in several places and get it to work under *some* circumstances.

As a side effect, added some more information to the plink strucutre.
This turned out to not be necessary for the solution I finally came
up with - it was part of an earlier, abandoned one. But, it's probably
good information to have around in the future.