First cut at porting our jail setup to linux vservers. Most of the changes are on the client side where I took mkjail and retargeted it to vservers (called it mkvserver.pl, clever eh?) in the linux directory. The real time sync was understanding how vservers work, how they boot how they die, how they handle signals, etc. Very interesting and very bizarre. Anyway, this first cut is done with the version 2.2 vserver code which does not virtualize the network stack or even the loopback device, so I pretty much ignored the experimental network and the host routine stuff. So, in your NS file you can now do this:
	set ns [new Simulator]
	set v0 [$ns node]
	set v1 [$ns node]

	tb-set-hardware $v0 pcvm
	tb-set-hardware $v1 pcvm
	tb-set-node-os $v0 FC-VSERVER
	tb-set-node-os $v1 FC-VSERVER

As you can see, I am using the osid to indicate jails vs
vservers. There are some small changes in assign_wrapper that use the
nextosid of the osid to map to the actual osid to install on the
hosting node. If you try to collocate a jail and a vserver assign will
refuse, cause we use features and desires for the osids. Sweet.

Oh, the ssh button in the web interface does not work yet cause the page
assumes that local virtnodes can bind to port 22 in each vserver, but
that will not work yet.