Add support to synch with and store plab nodegroups in our DB (plab uses them to key many different control actions, and we can't ignore them anymore).  The driving need in this case is to dist out rootballs based on the nodegroup so we can test plab4.2... but other ways to use this will likely arise.
Also expanded the notion of config attributes I added during the push to
support multiple PLCs.  Now, config attributes are just per-PLC or
per-slice; they are per-PLC, but also conditional across
slice,nodegroup,node tuples.  This gives a ton of flexibility for the
different actions we need to take based on what kind of node the node is
-- and plab typically seems to distinguish this based on nodegroups.