A few changes:
The valid time is now part of the rspec

Names are now solely for debugging, and are optional. UUIDs, which I
have just added, are the "real" way to uniquely identify nodes and
links.

Fix path to top.rnc, since I've moved this file

Use the stanard xsd:boolean type

A few re-fills