Make nodereboot respect the waittime arg, and wait 10 minutes for PLC.
Currently nodereboot in libreboot essentially ignores the waittime
arg because it forks and calls node_reboot to do the real work, but
doesn't pass on the waittime to it.  Fix this by adding a "-W"
option to node_reboot in order to specify the waittime.

Use this to extend the waittime for a PLC node to come up from 6 minutes to 10.