Add more wildcarding, mostly to support the test scripts.
You can now wildcard the node uuid and abd interface name in links:

        <node uuid="*"
              nickname="geni1"
              virtualization_type="emulab-vnode">
        </node>
        <node uuid="*"
              nickname="geni2"
              virtualization_type="emulab-vnode">
        </node>
        <link name="link0" nickname="link0">
         <linkendpoints nickname="destination_interface"
                   iface_name="*"
                   node_nickname="geni1"
                   node_uuid="*" />
         <linkendpoints nickname="source_interface"
                   iface_name="*"
                   node_nickname="geni2"
                   node_uuid="*" />
        </link>

So after the nodes are allocated, fill in the rest of the details in
the link part, including finding interfaces. The rspec is updated for
returned ticket.

The same approach can used for tunnel testing, but in that case you
allocate tickets to get the nodes, dig out the uuids, and then do an
UpdateSliver(), adding the tunnel specs.